refactor: deleting unnecessary network models
Refactor the series models to be unique and used across different screens
