+++ PopSeries/app/src/main/java/alifyz/com/popseries/adapter/HomeSeriesAdapter.kt
@@ -1,12 +1,11 @@
 package alifyz.com.popseries.adapter
 
 import alifyz.com.popseries.R
-import alifyz.com.popseries.model.PopularModel
-import alifyz.com.popseries.model.TopModel
+import alifyz.com.popseries.model.SeriesModel
 import alifyz.com.popseries.ui.DetailsActivity
 import android.content.Context
 import android.content.Intent
-import android.support.v4.content.ContextCompat
+import android.support.v4.content.ContextCompat.startActivity
 import android.support.v7.widget.CardView
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
@@ -17,10 +16,10 @@ import android.widget.TextView
 import com.bumptech.glide.Glide
 import com.google.gson.GsonBuilder
 
-class TopAdapter(val context: Context, val dataSet: TopModel) : RecyclerView.Adapter<TopAdapter.ViewHolder>() {
+class HomeSeriesAdapter(val context: Context, val dataSet: SeriesModel) : RecyclerView.Adapter<HomeSeriesAdapter.ViewHolder>() {
 
 
-    class ViewHolder(view: View) : RecyclerView.ViewHolder(view), View.OnClickListener {
+    inner class ViewHolder(view: View) : RecyclerView.ViewHolder(view), View.OnClickListener {
 
         val title = view.findViewById<TextView>(R.id.title)
         val year = view.findViewById<TextView>(R.id.year)
@@ -29,7 +28,14 @@ class TopAdapter(val context: Context, val dataSet: TopModel) : RecyclerView.Ada
                 .setOnClickListener(this)
 
         override fun onClick(v: View?) {
+            val seriesDetail = dataSet.results?.get(adapterPosition)
+            val builder = GsonBuilder()
+            val gson = builder.create()
 
+            val intent = Intent(context, DetailsActivity::class.java)
+            intent.putExtra("data", gson.toJson(seriesDetail))
+
+            startActivity(context, intent, null)
         }
     }
 
@@ -38,7 +44,7 @@ class TopAdapter(val context: Context, val dataSet: TopModel) : RecyclerView.Ada
     }
 
     override fun getItemCount(): Int {
-        return dataSet.results!!.size
+       return dataSet.results!!.size
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
