--- PopSeries/app/src/main/java/alifyz/com/popseries/network/NetworkHelper.kt
@@ -1,3 +1,0 @@
-package alifyz.com.popseries.network
-
-//TODO - Implement the RetrofitHelper Instance
